fix: Gets correct node for paths that start with the same string
CC-566
The getNodeForPath was checking if a node was starting with a string, instead of checking for equality. This resulted in a bug where if we had similar start of a path, it would return the first string it found. Now we fix that by checking if the string is equal. We also added a startWith "path + [" for the cases where the yaml path ends with array and we want to return that object itself.